I've made changes to the main method's flow.
Previously, the main method called the info method to display the instructions and then asked the user for their name. Now, the main method stores the quiz heading in a string variable called "quizHeading", which will be used to display the heading in other methods such as "Name", "Level", "Info". The updated method will call a new method called "Name" which returns the user's name, which is then stored in the "name" string variable. The "Level" method will ask the user for their level and store it in an integer variable called "level". The "Info" method will display the quiz instructions. Then "Quiz" method will be call to conduct the quiz, after "Quiz " is completed, the user's score will be stored in the "score" variable. The "Summary" method will display the quiz report, and the "QuizRedo" method will prompt the user to redo the quiz if they wish to do so.

I still have to work on the following :

Name method
Level Method
Info Method
Quiz Method
Summary method
QuizRedo Method